Revoke date of birth field validation
The date of birth validation will be implemented in the registration wizard instead, so validation of this field only happens for new users and not for existing users changing something in their profile. This was decided because there are many users without a DoB filled in and we can not fix this easily. Because users are not allowed to change their DoB, validating the DoB during edit-profile, would mean they could not edit their profile anymore.